Updated backend request update/delete to allow getting request by requestno

+++ backend/main.py
@@ -7,7 +7,6 @@ import traceback
 from flask import Flask, request, jsonify
 from firebase_admin import credentials, initialize_app, firestore
 from datetime import date
-import time
 
 # Initialize Flask app
 app = Flask(__name__)
@@ -18,7 +17,6 @@ default_app = initialize_app(cred)
 db = firestore.client()
 users_ref = db.collection('users')
 requests_ref = db.collection('requests')
-requestno_ref = db.collection('requestno')
 
 error_500 = """<!doctype html>
 <html lang=en>
@@ -38,20 +36,14 @@ def add_user():
     """
     try:
         # should include username, password, email, and permissions
-        if request.form.get('id') is None:
-            # doc_id = users_ref.add(request.form)
-            # return jsonify({"success": True, "id": doc_id[1].id}), 201
-            return jsonify({"success": False, "reason": "No UID provided."}), 400
-        else:
-            users_ref.document(request.form['id']).set(request.form)
-            return jsonify({"success": True, "id": request.form['id']}), 201
+        doc_id = users_ref.add(request.form)
+        return jsonify({"success": True, "id": doc_id[1].id}), 201
     except Exception as e:
         print(f"An Error Occurred: {e}")
         return error_500, 500
 
 
 @app.route('/users', methods=['GET'])
-@app.route('/users/list', methods=['GET'])
 def get_user():
     """
         get_user() : Fetches documents from Firestore collection as JSON.
@@ -60,17 +52,16 @@ def get_user():
         # Check if ID was passed to URL query
         user_id = request.args.get('id')
         username = request.args.get('username')
-        if user_id is None and username is not None:
+        if user_id is None:
             for doc in users_ref.stream():
                 if doc.get('username') == username:
                     user_id = doc.id
-        if user_id is not None:
+        if user_id:
             user = users_ref.document(user_id).get()
             return jsonify(user.to_dict()), 200
         else:
-            all_users = [doc.to_dict() for doc in users_ref.stream()]
-            return jsonify(all_users), 200
-            # return jsonify("No user found matching given parameters."), 404
+            # all_users = [doc.to_dict() for doc in users_ref.stream()]
+            return jsonify("No user found matching given parameters."), 404
     except Exception as e:
         print(f"An Error Occurred: {e}")
         return error_500, 500
@@ -130,9 +121,8 @@ def create_hygiene_request():
         # later change to sending user in request instead of demographic information needing to be resubmitted 24/7
         form_data = request.form.to_dict()
         form_data['date'] = date.today().strftime("%m/%d/%Y")  # MM/DD/YYYY
-        form_data['timestamp'] = str(time.time())
         form_data['type'] = 'hygiene'
-        form_data['status'] = 'submitted'
+        form_data['status'] = 'received'
         doc_id = requests_ref.add(form_data)
         return jsonify({"success": True, "id": doc_id[1].id}), 201
     except Exception as e:
@@ -149,9 +139,8 @@ def create_clothing_request():
         # later change to sending user in request instead of demographic information needing to be resubmitted 24/7
         form_data = request.form.to_dict()
         form_data['date'] = date.today().strftime("%m/%d/%Y")  # MM/DD/YYYY
-        form_data['timestamp'] = str(time.time())
         form_data['type'] = 'clothing'
-        form_data['status'] = 'submitted'
+        form_data['status'] = 'received'
         doc_id = requests_ref.add(form_data)
         return jsonify({"success": True, "id": doc_id[1].id}), 201
     except Exception as e:
@@ -165,21 +154,8 @@ def list_clothing_requests():
     try:
         # Check for ID in request args
         document_id = request.args.get('id')
-        uid = request.args.get('uid')
-        if uid is not None:
+        if document_id is None:
             all_requests = [doc.to_dict() for doc in requests_ref.stream()]
-            all_requests = sorted(all_requests, key=lambda x: x['timestamp'])
-            removed_documents = []
-            for i in range(len(all_requests)):
-                request_document = all_requests[i]
-                if request_document['type'].lower() != 'clothing' or request_document['uid'] != uid:
-                    removed_documents.append(request_document)
-            for document in removed_documents:
-                all_requests.remove(document)
-            return jsonify(all_requests), 200
-        elif document_id is None:
-            all_requests = [doc.to_dict() for doc in requests_ref.stream()]
-            all_requests = sorted(all_requests, key=lambda x: x['timestamp'])
             removed_documents = []
             for i in range(len(all_requests)):
                 request_document = all_requests[i]
@@ -201,21 +177,8 @@ def list_hygiene_requests():
     try:
         # Check for ID in request args
         document_id = request.args.get('id')
-        uid = request.args.get('uid')
-        if uid is not None:
+        if document_id is None:
             all_requests = [doc.to_dict() for doc in requests_ref.stream()]
-            all_requests = sorted(all_requests, key=lambda x: x['timestamp'])
-            removed_documents = []
-            for i in range(len(all_requests)):
-                request_document = all_requests[i]
-                if request_document['type'].lower() != 'hygiene' or request_document['uid'] != uid:
-                    removed_documents.append(request_document)
-            for document in removed_documents:
-                all_requests.remove(document)
-            return jsonify(all_requests), 200
-        elif document_id is None:
-            all_requests = [doc.to_dict() for doc in requests_ref.stream()]
-            all_requests = sorted(all_requests, key=lambda x: x['timestamp'])
             removed_documents = []
             for i in range(len(all_requests)):
                 request_document = all_requests[i]
@@ -237,21 +200,8 @@ def list_all_requests():
     try:
         # Check for ID in request args
         document_id = request.args.get('id')
-        uid = request.args.get('uid')
-        if uid is not None:
+        if document_id is None:
             all_requests = [doc.to_dict() for doc in requests_ref.stream()]
-            all_requests = sorted(all_requests, key=lambda x: x['timestamp'])
-            removed_documents = []
-            for i in range(len(all_requests)):
-                request_document = all_requests[i]
-                if request_document['uid'] != uid:
-                    removed_documents.append(request_document)
-            for document in removed_documents:
-                all_requests.remove(document)
-            return jsonify(all_requests), 200
-        elif document_id is None:
-            all_requests = [doc.to_dict() for doc in requests_ref.stream()]
-            all_requests = sorted(all_requests, key=lambda x: x['timestamp'])
             return jsonify(all_requests), 200
         else:
             return jsonify(requests_ref.document(document_id).get().to_dict()), 200
@@ -266,8 +216,17 @@ def remove_request():
     try:
         # Check for ID in URL query
         document_id = request.args.get('id')
-        requests_ref.document(document_id).delete()
-        return jsonify({"success": True}), 200
+        requestno = request.args.get('requestno')
+        if document_id is None and requestno is not None:
+            for doc in requests_ref.stream():
+                if doc.get('requestno') == requestno:
+                    document_id = doc.id
+                    break
+        if document_id is not None:
+            requests_ref.document(document_id).delete()
+            return jsonify({"success": True}), 200
+        else:
+            return jsonify("No request found matching given parameters."), 404
     except Exception as e:
         print(f"An Error Occurred: {e}")
         return error_500, 500
@@ -278,24 +237,18 @@ def remove_request():
 def update_request():
     try:
         document_id = request.args.get('id')
-        requests_ref.document(document_id).update(request.form)
-        return jsonify({"success": True}), 200
-    except Exception as e:
-        print(f"An Error Occurred: {e}")
-        return error_500, 500
+        requestno = request.args.get('requestno')
+        if document_id is None and requestno is not None:
+            for doc in requests_ref.stream():
+                if doc.get('requestno') == requestno:
+                    document_id = doc.id
+                    break
+        if document_id is not None:
+            requests_ref.document(document_id).update(request.form)
+            return jsonify({"success": True}), 200
+        else:
+            return jsonify("No request found matching given parameters."), 404
 
-
-@app.route('/requestno', methods=['GET'])
-def get_next_requestno():
-    try:
-        document_id = 'requestnum'
-        next_id = requestno_ref.document(document_id).get().to_dict()
-        next_id = next_id['nextID']
-        update_id = dict()
-        update_id['nextID'] = next_id + 1
-        requestno_ref.document('requestnum').update(update_id)
-
-        return jsonify({'id': next_id}), 200
     except Exception as e:
         print(f"An Error Occurred: {e}")
         return error_500, 500
